test_pulsar/requirements.txt
Django==4.1.4
django-filter==22.1
djangorestframework==3.14.0
Pillow==9.3.0
